# docs/swagger.yaml
basePath: /
definitions:
...
  rest.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
  weekly_plan.WeeklyPlanItemDTO:
    properties:
      budgetItemId:
        type: integer
      color:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      position:
        type: integer
      weeklyDuration:
        type: integer
      weeklyOccurrences:
        type: integer
    type: object
host: localhost:8181
info:
  contact: {}
  description: Time planning and tracking API
  title: Klokku API
  version: "0.9"
paths:
...
  /api/weeklyplan:
    delete:
      description: Reset all weekly plan items to budget plan values for a specific
        week
      parameters:
      - description: Date in RFC3339 format (can be any day of the week)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/weekly_plan.WeeklyPlanItemDTO'
            type: array
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Reset entire week
      tags:
      - WeeklyPlan
    get:
      description: Retrieve all items for a specific week
      parameters:
      - description: Date in RFC3339 format (can be any day of the week)
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/weekly_plan.WeeklyPlanItemDTO'
            type: array
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Get weekly plan items
      tags:
      - WeeklyPlan
  /api/weeklyplan/item:
    put:
      consumes:
      - application/json
      description: Update the duration and notes for a weekly plan item
      parameters:
      - description: Date in RFC3339 format (can be any day of the week)
        in: query
        name: date
        required: true
        type: string
      - description: Item update details
        in: body
        name: item
        required: true
        schema:
          properties:
            budgetItemId:
              type: integer
            id:
              type: integer
            notes:
              type: string
            weeklyDuration:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/weekly_plan.WeeklyPlanItemDTO'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: User not found
          schema:
            type: string
        "404":
          description: Item Not Found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Update a weekly plan item
      tags:
      - WeeklyPlan
  /api/weeklyplan/item/{itemId}:
    delete:
      description: Reset a weekly plan item to its original budget plan values
      parameters:
      - description: Weekly Plan Item ID
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/weekly_plan.WeeklyPlanItemDTO'
        "400":
          description: Invalid itemId
          schema:
            $ref: '#/definitions/rest.ErrorResponse'
        "403":
          description: User not found
          schema:
            type: string
      security:
      - XUserId: []
      summary: Reset a weekly plan item
      tags:
      - WeeklyPlan
securityDefinitions:
  XUserId:
    description: User ID header required for authentication
    in: header
    name: X-User-Id
    type: apiKey
swagger: "2.0"